// src/app/core/api/model/createCampaignOfferDto.ts
/**
 * Trendors
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferLineDto } from './offerLineDto';


export interface CreateCampaignOfferDto { 
    /**
     * Parent CampaignInfluencer ID (must be invited)
     */
    campaignInfluencerId: number;
    proposed_by: CreateCampaignOfferDto.ProposedByEnum;
    message?: string;
    expires_at?: string;
    lines?: Array<OfferLineDto>;
}
export namespace CreateCampaignOfferDto {
    export const ProposedByEnum = {
        Brand: 'brand',
        Influencer: 'influencer',
    } as const;
    export type ProposedByEnum = typeof ProposedByEnum[keyof typeof ProposedByEnum];
}